avfilter/scale_eval: Use 64bit for factor_w/h
Avoids truncation and overflows

Fixes: #YWH-PGM40646-14
Found-by: An0n99X
(cherry picked from commit 5bbc8f828e85a4a2ded42f7d46b2ce8e2657433d)

+++ libavfilter/scale_eval.c
@@ -124,7 +124,7 @@ int ff_scale_adjust_dimensions(AVFilterLink *inlink,
     int force_original_aspect_ratio, int force_divisible_by)
 {
     int64_t w, h;
-    int factor_w, factor_h;
+    int64_t factor_w, factor_h;
 
     w = *ret_w;
     h = *ret_h;
